Blacklist keystonemiddleware 4.19
The 4.19 release of keystonemiddleware contains a commit[0] which
starts making use of the oslo.cache library and specifically some
internal module which interface with memcache. However
python-memcache is an optional backend of the library, and
therefore not installed by default. Additionally, when
keystonemiddleware is not configured for caching, those modules
are not needed. So python-memcache can't be a hard dependency of
keystonemiddleware, but the code treats it as such. This is tracked
at bug [1].

[2] merged which enables lazy loading for oslo.cache, but that
won't be included until the next release, so we need to blacklist
4.19.

0. https://github.com/openstack/keystonemiddleware/commit/9d8e2836fe7fca186e0380d8a532540ff5cc5215
1. https://bugs.launchpad.net/keystonemiddleware/+bug/1737115
2. https://github.com/openstack/keystonemiddleware/commit/35fa0e1da1b96a40aa2aa6dd2be378e4155a22ba